+++ src/Component/Header.js
@@ -13,7 +13,7 @@ function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
 }
 
-export default function Header() {
+export default function Header(props) {
   return (
     <Disclosure as="nav" className="bg-gray-800">
       {({ open }) => (
@@ -98,6 +98,7 @@ export default function Header() {
           </Disclosure.Panel>
         </>
       )}
+      {props.children}
     </Disclosure>
   );
 }
